bluetooth: hci_bcm: Fix shutdown GPIO polarity
The semantic of the shutdown GPIO is used backward in the driver: while the
GPIO is called shutdown, it is enabled when the device is powered up.

This doesn't make much sense, since that would imply that we are shutting
down the device, while the code path clearly indicates that we power it up.

Invert the logic to set it low when we power the chip, and high when we
disable it, which is the sound thing to do. We'll also need to fix the
ACPI GPIOs polarity so that things don't get broken.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>