HACK: media: vb2: Make video_device available when allocating buffers
In order to track down buffer allocations through cgroup, we need to
come up with a generic way to register all these buffer allocations.

__vb2_buf_mem_alloc and __vb2_buf_mem_free seem to be like the best
place: they are central and used by every user, and yet and shared
between all videobuf2 implementations.

However, they don't have access to the video_device we want to track
allocations for.

We discussed it with Laurent recently, and his suggestion was to make a
much larger rework to the v4l2 framework in order to support this.

It's definitely something we can work on, however it goes against the
objective to make the device cgroup work progress.

So, for now, and with the objective to show that the device cgroup API
is generic enough, let's make a quick hack to provide the video_device
to __vb2_buf_mem_alloc, which is probably all kinds of broken.

Signed-off-by: Maxime Ripard <mripard@kernel.org>